Randomized Spawning & Edits to default colors
Made a randomized ball spawning system. Which randomizes its spawn position and color. And made it so when no theme is equipped you get the default colors. One problem with randomized spawning system is that balls spawn on top of each other and/or in each other, which cause an explosion or balls moving to the wrong corner by themselves.