Pause Menu, Changed how you exit game, added a Slow Motion ability.
There is now a Pause Menu during the levels so you can leave the game or go to the main menu. Instead of pressing esc in the main menu to exit the game there is now a button. I added a Slow Motion ability the player can use to their advantage once all levels have been completed.  As well as a simple tutorial on how to use it in the HOw to play Page and when you use it for the first time.